requirements.txt
pkutils
matplotlib
pandas
autogluon.tabular==0.2.0
flask
gunicorn
